Update how tests are executed
This PR aims to improve the way tests are executed. Specifically, it
makes it so that fixtures are loaded and unloaded by the TestCase.
Issues with the fixture loading was reported in #36.

Currently, fixtures are loaded via the `loaddata` command and we rely on
the transaction to clear it at the end of the test. It turns out that
this isn't a good way of doing it and there are a lot of bugs associated
with sequences and primary keys.

So instead, we'll rely on Django's TestCase (whichever one is used) to
load and clear the fixtures. As a bonus, this also lets us use the
`reset_sequences` feature.